Moved CivOne.Sprites to CivOne.Graphics.Sprites
