+++ src/Graphics/Sprites/ISprite.cs
@@ -7,17 +7,12 @@
 // You should have received a copy of the CC0 legalcode along with this
 // work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 
-using CivOne.Graphics;
+using CivOne.IO;
 
-namespace CivOne.Sprites
+namespace CivOne.Graphics.Sprites
 {
-	public static class SpriteExtensions
+	public interface ISprite
 	{
-		public static IBitmap ToBitmap(this ISprite sprite)
-		{
-			if (sprite == null)
-				return null;
-			return new Picture(sprite.Bitmap, Common.DefaultPalette);
-		}
+		Bytemap Bitmap { get; }
 	}
 }
