+++ src/Graphics/Sprites/Patterns.cs
@@ -7,9 +7,7 @@
 // You should have received a copy of the CC0 legalcode along with this
 // work. If not, see <http://creativecommons.org/publicdomain/zero/1.0/>.
 
-using CivOne.Graphics;
-
-namespace CivOne.Sprites
+namespace CivOne.Graphics.Sprites
 {
 	public static class Patterns
 	{
