+++ src/Screens/CityChooseProduction.cs
@@ -13,7 +13,7 @@ using System.Linq;
 using CivOne.Buildings;
 using CivOne.Enums;
 using CivOne.Graphics;
-using CivOne.Sprites;
+using CivOne.Graphics.Sprites;
 using CivOne.Units;
 using CivOne.UserInterface;
 using CivOne.Wonders;
