+++ src/Screens/CityManagerPanels/CityFoodStorage.cs
@@ -11,7 +11,7 @@ using System.Linq;
 using CivOne.Buildings;
 using CivOne.Enums;
 using CivOne.Graphics;
-using CivOne.Sprites;
+using CivOne.Graphics.Sprites;
 
 namespace CivOne.Screens.CityManagerPanels
 {
