+++ src/Screens/CityManagerPanels/CityInfo.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using CivOne.Enums;
 using CivOne.Events;
 using CivOne.Graphics;
-using CivOne.Sprites;
+using CivOne.Graphics.Sprites;
 using CivOne.Units;
 
 namespace CivOne.Screens.CityManagerPanels
