+++ src/Screens/Debug/SpawnUnit.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using CivOne.Enums;
 using CivOne.Events;
 using CivOne.Graphics;
-using CivOne.Sprites;
+using CivOne.Graphics.Sprites;
 using CivOne.Tiles;
 using CivOne.Units;
 using CivOne.UserInterface;
