+++ src/Screens/Dialogs/BaseDialog.cs
@@ -12,7 +12,7 @@ using System.Linq;
 using CivOne.Enums;
 using CivOne.Events;
 using CivOne.Graphics;
-using CivOne.Sprites;
+using CivOne.Graphics.Sprites;
 
 namespace CivOne.Screens.Dialogs
 {
