+++ src/Screens/Newspaper.cs
@@ -13,7 +13,7 @@ using CivOne.Enums;
 using CivOne.Events;
 using CivOne.IO;
 using CivOne.Graphics;
-using CivOne.Sprites;
+using CivOne.Graphics.Sprites;
 
 namespace CivOne.Screens
 {
