+++ src/Screens/Overlay.cs
@@ -13,7 +13,7 @@ using System.Linq;
 using CivOne.Enums;
 using CivOne.Events;
 using CivOne.Graphics;
-using CivOne.Sprites;
+using CivOne.Graphics.Sprites;
 using CivOne.Tiles;
 using CivOne.Units;
